some small fixes and docs

+++ src/helper/decorators.ts
@@ -1,5 +1,6 @@
 import * as NodeCache from 'node-cache';
 import ApiError from './ApiError';
+import { HttpStatus } from '@nestjs/common';
 
 const caches = new Map<string, NodeCache>();
 
@@ -15,48 +16,45 @@ const caches = new Map<string, NodeCache>();
  * Limitations:
  * - Caching is in-memory and resets on application restart.
  */
-export function CacheResult(
-    ttl: number = 60,
-    cacheId: string = "default"
-  ): MethodDecorator {
-    if (!caches.has(cacheId)) {
-      caches.set(cacheId, new NodeCache({ stdTTL: ttl }));
+export function CacheResult<T>(
+  ttl: number = 60,
+  cacheId: string = "default"
+): MethodDecorator {
+  if (!caches.has(cacheId)) 
+    caches.set(cacheId, new NodeCache({ stdTTL: ttl }));
+  
+  const cache = caches.get(cacheId)!;
+
+  return function (
+    _: Object,
+    propertyKey: string | symbol,
+    descriptor: PropertyDescriptor
+  ): void {
+    const originalMethod = descriptor.value as (...args: unknown[]) => T;
+
+    if (typeof originalMethod !== "function") {
+      throw new ApiError(HttpStatus.INTERNAL_SERVER_ERROR, `Descriptor value for method ${String(propertyKey)} is not a function.`);
     }
-    const cache = caches.get(cacheId)!;
-  
-    return function (
-      _: Object,
-      propertyKey: string | symbol,
-      descriptor: PropertyDescriptor
-    ): void {
-      const originalMethod = descriptor.value;
-  
-      if (typeof originalMethod !== "function") {
-        throw new ApiError(500, `Descriptor value for method ${String(propertyKey)} is not a function.`);
-      }
-  
-      descriptor.value = function (...args: unknown[]): unknown {
-        const cacheKey = `${String(propertyKey)}-${JSON.stringify(args)}`;
-        const cachedResult = cache.get<unknown>(cacheKey);
-  
-        if (cachedResult !== undefined) {
-          return cachedResult;
-        }
-  
-        const result = originalMethod.apply(this, args);
-  
-        if (result instanceof Promise) {
-          return result.then((resolvedResult) => {
-            cache.set(cacheKey, resolvedResult);
-            return resolvedResult;
-          });
-        }
-  
-        cache.set(cacheKey, result);
-        return result;
-      };
+
+    descriptor.value = function (...args: unknown[]): T | Promise<T> {
+      const cacheKey = `${String(propertyKey)}-${JSON.stringify(args)}`;
+      const cachedResult = cache.get<T>(cacheKey);
+
+      if (cachedResult !== undefined) return cachedResult;
+
+      const result = originalMethod.apply(this, args);
+
+      if (result instanceof Promise) 
+        return result.then((resolvedResult: T) => {
+          cache.set(cacheKey, resolvedResult);
+          return resolvedResult;
+        });
+
+      cache.set(cacheKey, result);
+      return result;
     };
-  }
+  };
+}
 
 /**
  * ClearCache is a method decorator that clears all cached results for a specific `cacheId` 
@@ -77,27 +75,23 @@ export function ClearCache(cacheId: string = 'default'): MethodDecorator {
       propertyKey: string | symbol,
       descriptor: TypedPropertyDescriptor<T>
     ): void {
-      const originalMethod = descriptor.value;
+      const originalMethod = descriptor.value as (...args: unknown[]) => T;
   
-      if (typeof originalMethod !== 'function') {
-        throw new ApiError(500, `Descriptor value is not a function for method ${String(propertyKey)}`);
-      }
+      if (typeof originalMethod !== 'function') 
+        throw new ApiError(HttpStatus.INTERNAL_SERVER_ERROR, `Descriptor value is not a function for method ${String(propertyKey)}`);
+      
   
-      descriptor.value = function (...args: unknown[]) {
+      descriptor.value = function (...args: unknown[]): T | Promise<T> {
         const result = originalMethod.apply(this, args);
   
-        if (result instanceof Promise) {
-          return result.then((resolvedResult) => {
-            if (caches.has(cacheId)) {
-              caches.get(cacheId)!.flushAll();
-            }
+        if (result instanceof Promise) 
+          return result.then((resolvedResult: T) => {
+            if (caches.has(cacheId)) caches.get(cacheId)!.flushAll();
+            
             return resolvedResult;
           });
-        }
-  
-        if (caches.has(cacheId)) {
-          caches.get(cacheId)!.flushAll();
-        }
+        
+        if (caches.has(cacheId)) caches.get(cacheId)!.flushAll();
   
         return result;
       } as T;
